Working MPI Implementation

+++ src/matrix/mat_factory.h
@@ -9,6 +9,6 @@
 
 mat_t *mat_factory_init_empty(long xSize, long ySize);
 
-mat_t *mat_factory_init_seeded(long xSize, long ySize);
+mat_t *mat_factory_init_seeded(int xSize, int ySize);
 
 #endif //PARALLEL_COMPUTATION_CW1_MATRIXFACTORY_H
